[ADD] Added service worker for PWA
